+++ applications/FloatInstrumentation/FloatInterval.cpp
@@ -42,7 +42,7 @@
 
 namespace NumericalDomains { namespace DDoubleInterval {
 
-template <> int tfinite(long double val) { return finite(val); }
+template <> int tfinite(long double val) { return finite((double) val); }
 template <> int tfinite(double val) { return finite(val); }
 template <> int tfinite(float val) { return finite(val); }
